# docs/architecture/api-spec.yaml
openapi: 3.1.0
info:
  title: JMW Agent Facts Server API
  version: "1.0.0"
  description: >
    Iteration-2 API for the JMW Agent Facts Server. Two principal types:
    agents (Bearer API key) on /api/v1/agent/*, and users (session cookie) on
    all other endpoints. All list endpoints use keyset pagination. Errors use a
    uniform envelope. See api-spec.md and conventions.md.
servers:
  - url: /api/v1
    description: Versioned base path (single version this iteration)
components:
  securitySchemes:
    agentApiKey:
      type: http
      scheme: bearer
      description: Agent API key issued at registration; only its hash is stored.
    userSession:
      type: apiKey
      in: cookie
      name: session
      description: httpOnly Secure SameSite=Strict session cookie.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: snake_case domain code }
            message: { type: string }
...
    AgentConfigRequest:
      type: object
      required: [heartbeat_interval_secs, discovery_interval_secs, inventory_interval_secs]
      properties:
        heartbeat_interval_secs: { type: integer }
        discovery_interval_secs: { type: integer }
        inventory_interval_secs: { type: integer }
    CollectorToggleRequest:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
    CollectorIntervalRequest:
      type: object
      required: [interval_secs]
      properties:
        interval_secs: { type: integer }
...
  responses:
...
    Forbidden:
      description: RBAC or approval gate
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
paths:
...
  /admin/agents/{id}/config:
    put:
      summary: Update agent collection intervals (admin)
      security: [{ userSession: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AgentConfigRequest' } } }
      responses:
        '200': { description: Config updated; delivered on next heartbeat }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/agents/{id}/collectors/{collector_name}/toggle:
    post:
      summary: Enable/disable a specific collector for an agent (admin)
      security: [{ userSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: collector_name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CollectorToggleRequest' } } }
      responses:
        '200': { description: Collector toggled }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Unknown collector, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/agents/{id}/collectors/{collector_name}/interval:
    put:
      summary: Set a per-collector frequency override (admin)
      security: [{ userSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: collector_name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CollectorIntervalRequest' } } }
      responses:
        '200': { description: Interval override set }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Unknown collector or bad interval, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...